Create WaterAddButton component
